refactor out QubitQNode, CVQnode
