+++ pennylane/qnode_new/qnode.py
@@ -21,7 +21,6 @@ import inspect
 import itertools
 
 import numpy as np
-from scipy import linalg
 
 import pennylane as qml
 from pennylane.measure import var
@@ -560,202 +559,3 @@ class QNode:
         self.device.reset()
         ret = self.device.execute(self.circuit.operations, obs, self.circuit.variable_deps)
         return ret
-
-    def _construct_metric_tensor(self, *, diag_approx=False):
-        """Construct metric tensor subcircuits for qubit circuits.
-
-        Constructs a set of quantum circuits for computing a block-diagonal approximation of the
-        Fubini-Study metric tensor on the parameter space of the variational circuit represented
-        by the QNode, using the Quantum Geometric Tensor.
-
-        If the parameter appears in a gate :math:`G`, the subcircuit contains
-        all gates which precede :math:`G`, and :math:`G` is replaced by the variance
-        value of its generator.
-
-        Args:
-            diag_approx (bool): iff True, use the diagonal approximation
-
-        Raises:
-            QuantumFunctionError: if a metric tensor cannot be generated because no generator
-                was defined
-
-        """
-        # pylint: disable=too-many-statements, too-many-branches
-
-        self._metric_tensor_subcircuits = {}
-        for queue, curr_ops, param_idx, _ in self.circuit.iterate_layers():
-            obs = []
-            scale = []
-
-            Ki_matrices = []
-            KiKj_matrices = []
-            Ki_ev = []
-            KiKj_ev = []
-            V = None
-
-            # for each operation in the layer, get the generator and convert it to a variance
-            for n, op in enumerate(curr_ops):
-                gen, s = op.generator
-                w = op.wires
-
-                if gen is None:
-                    raise QuantumFunctionError(
-                        "Can't generate metric tensor, operation {}"
-                        "has no defined generator".format(op)
-                    )
-
-                # get the observable corresponding to the generator of the current operation
-                if isinstance(gen, np.ndarray):
-                    # generator is a Hermitian matrix
-                    variance = var(qml.Hermitian(gen, w, do_queue=False))
-
-                    if not diag_approx:
-                        Ki_matrices.append((n, expand(gen, w, self.num_wires)))
-
-                elif issubclass(gen, Observable):
-                    # generator is an existing PennyLane operation
-                    variance = var(gen(w, do_queue=False))
-
-                    if not diag_approx:
-                        if issubclass(gen, qml.PauliX):
-                            mat = np.array([[0, 1], [1, 0]])
-                        elif issubclass(gen, qml.PauliY):
-                            mat = np.array([[0, -1j], [1j, 0]])
-                        elif issubclass(gen, qml.PauliZ):
-                            mat = np.array([[1, 0], [0, -1]])
-
-                        Ki_matrices.append((n, expand(mat, w, self.num_wires)))
-
-                else:
-                    raise QuantumFunctionError(
-                        "Can't generate metric tensor, generator {}"
-                        "has no corresponding observable".format(gen)
-                    )
-
-                obs.append(variance)
-                scale.append(s)
-
-            if not diag_approx:
-                # In order to compute the block diagonal portion of the metric tensor,
-                # we need to compute 'second order' <psi|K_i K_j|psi> terms.
-
-                for i, j in itertools.product(range(len(Ki_matrices)), repeat=2):
-                    # compute the matrices representing all K_i K_j terms
-                    obs1 = Ki_matrices[i]
-                    obs2 = Ki_matrices[j]
-                    KiKj_matrices.append(((obs1[0], obs2[0]), obs1[1] @ obs2[1]))
-
-                V = np.identity(2 ** self.num_wires, dtype=np.complex128)
-
-                # generate the unitary operation to rotate to
-                # the shared eigenbasis of all observables
-                for _, term in Ki_matrices:
-                    _, S = linalg.eigh(V.conj().T @ term @ V)
-                    V = np.round(V @ S, 15)
-
-                V = V.conj().T
-
-                # calculate the eigenvalues for
-                # each observable in the shared eigenbasis
-                for idx, term in Ki_matrices:
-                    eigs = np.diag(V @ term @ V.conj().T).real
-                    Ki_ev.append((idx, eigs))
-
-                for idx, term in KiKj_matrices:
-                    eigs = np.diag(V @ term @ V.conj().T).real
-                    KiKj_ev.append((idx, eigs))
-
-            self._metric_tensor_subcircuits[param_idx] = {
-                "queue": queue,
-                "observable": obs,
-                "Ki_expectations": Ki_ev,
-                "KiKj_expectations": KiKj_ev,
-                "eigenbasis_matrix": V,
-                "result": None,
-                "scale": scale,
-            }
-
-    def metric_tensor(self, args, kwargs=None, *, diag_approx=False, only_construct=False):
-        """Evaluate the value of the metric tensor.
-
-        Args:
-            args (tuple[Any]): positional (differentiable) arguments
-            kwargs (dict[str, Any]): auxiliary arguments
-            diag_approx (bool): iff True, use the diagonal approximation
-            only_construct (bool): Iff True, construct the circuits used for computing
-                the metric tensor but do not execute them, and return None.
-
-        Returns:
-            array[float]: metric tensor
-        """
-        kwargs = kwargs or {}
-        kwargs = self._default_args(kwargs)
-
-        if self.circuit is None or self.mutable:
-            # construct the circuit
-            self._construct(args, kwargs)
-
-        if self._metric_tensor_subcircuits is None:
-            self._construct_metric_tensor(diag_approx=diag_approx)
-
-        if only_construct:
-            return None
-
-        # temporarily store the parameter values in the Variable class
-        self._set_variables(args, kwargs)
-
-        tensor = np.zeros([self.num_variables, self.num_variables])
-
-        # execute constructed metric tensor subcircuits
-        for params, circuit in self._metric_tensor_subcircuits.items():
-            self.device.reset()
-
-            s = np.array(circuit["scale"])
-            V = circuit["eigenbasis_matrix"]
-
-            if not diag_approx:
-                # block diagonal approximation
-
-                unitary_op = qml.QubitUnitary(V, wires=list(range(self.num_wires)), do_queue=False)
-                self.device.execute(circuit["queue"] + [unitary_op], circuit["observable"])
-                probs = list(self.device.probability().values())
-
-                first_order_ev = np.zeros([len(params)])
-                second_order_ev = np.zeros([len(params), len(params)])
-
-                for idx, ev in circuit["Ki_expectations"]:
-                    first_order_ev[idx] = ev @ probs
-
-                for idx, ev in circuit["KiKj_expectations"]:
-                    # idx is a 2-tuple (i, j), representing
-                    # generators K_i, K_j
-                    second_order_ev[idx] = ev @ probs
-
-                    # since K_i and K_j are assumed to commute,
-                    # <psi|K_j K_i|psi> = <psi|K_i K_j|psi>,
-                    # and thus the matrix of second-order expectations
-                    # is symmetric
-                    second_order_ev[idx[1], idx[0]] = second_order_ev[idx]
-
-                g = np.zeros([len(params), len(params)])
-
-                for i, j in itertools.product(range(len(params)), repeat=2):
-                    g[i, j] = (
-                        s[i]
-                        * s[j]
-                        * (second_order_ev[i, j] - first_order_ev[i] * first_order_ev[j])
-                    )
-
-                row = np.array(params).reshape(-1, 1)
-                col = np.array(params).reshape(1, -1)
-                circuit["result"] = np.diag(g)
-                tensor[row, col] = g
-
-            else:
-                # diagonal approximation
-                circuit["result"] = s ** 2 * self.device.execute(
-                    circuit["queue"], circuit["observable"]
-                )
-                tensor[np.array(params), np.array(params)] = circuit["result"]
-
-        return tensor
